cont : move fit params implementation to libcommon
